backend/requirements.txt
django>=5.0
djangorestframework
psycopg2-binary
weasyprint
django-cors-headers
Pillow
python-decouple
gunicorn
whitenoise
